requirements.txt
# FastAPI框架及相关依赖
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
pydantic==1.10.7
python-multipart==0.0.6
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.1.0

# 认证相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 二维码生成
qrcode==7.4.2
pillow==9.5.0

# HTTP客户端
requests==2.28.2

# 国际化支持
python-i18n==0.3.9

# 邮件支持
aiosmtplib==2.0.1
